# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API App - Clinica",
        "version": "1.0"
    },
    "paths": {
...
        "/api/show/objetivos": {
            "get": {
                "tags": [
                    "Objetivos"
                ],
                "summary": "Mostrando los objetivos disponibles",
                "operationId": "showObjetivos",
                "responses": {
                    "200": {
                        "description": "Showing aviable Objetivos"
                    },
                    "403": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/show/objetivo/{objetivo}": {
            "get": {
                "tags": [
                    "Objetivos"
                ],
                "summary": "Mostrando el objetivo",
                "operationId": "showObjetivo",
                "parameters": [
                    {
                        "name": "objetivo",
                        "in": "path",
                        "description": "Id of the Objetivo",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Showing specific Objetivo"
                    },
                    "403": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/create/objetivo": {
            "post": {
                "tags": [
                    "Objetivos"
                ],
                "summary": "Crear objetivo",
                "operationId": "createObjetivo",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Objetivo created successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
        "/api/update/objetivo/{objetivo}": {
            "post": {
                "tags": [
                    "Objetivos"
                ],
                "summary": "Actualizar Objetivo",
                "operationId": "updateObjetivo",
                "parameters": [
                    {
                        "name": "objetivo",
                        "in": "path",
                        "description": "Id of the Objetivo",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Objetivo updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
        "/api/delete/objetivo/{objetivo}": {
            "get": {
                "tags": [
                    "Objetivos"
                ],
                "summary": "Eliminando el objetivo",
                "operationId": "deleteObjetivo",
                "parameters": [
                    {
                        "name": "objetivo",
                        "in": "path",
                        "description": "Id of the Objetivo",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Objetivo deleted successfully"
                    },
                    "403": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}